+++ core/src/main/java/org/cache2k/core/InternalCacheInfo.java
@@ -28,9 +28,9 @@ public interface InternalCacheInfo {
 
   String getImplementation();
 
-  int getSize();
+  long getSize();
 
-  int getMaxSize();
+  long getMaxSize();
 
   long getLoadButHitCnt();
